Исправил рейтинг

+++ main-service/src/main/java/ru/practicum/event/mapper/EventMapper.java
@@ -22,7 +22,7 @@ import java.time.LocalDateTime;
 public interface EventMapper {
     @Mapping(target = "rating", source = "rating")
     @Mapping(target = "views", source = "views")
-    EventFullDto toEventFullDto(final Event event, final int rating, final long views);
+    EventFullDto toEventFullDto(final Event event, final long rating, final long views);
 
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "publishedOn", ignore = true)
@@ -38,5 +38,5 @@ public interface EventMapper {
 
     @Mapping(target = "rating", source = "rating")
     @Mapping(target = "views", source = "views")
-    EventShortDto toEventShortDto(final Event event,final int rating, final long views);
+    EventShortDto toEventShortDto(final Event event,final long rating, final long views);
 }
